backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask-pymongo
flask-cors
python-dotenv
google-generativeai
langchain
langchain-google-genai
pydantic
